add user id

+++ controllers/videos.js
@@ -1,24 +1,27 @@
 const express = require('express');
 const router = express.Router();
 const db = require('../models');
+const auth = require('../middleware/auth');
 
-router.get('/', (req, res) => {
-	const userId = '5e1e6c078bd03303a274ba33';
-	db.Video.find({ userId })
+//Get all videos
+router.get('/', auth, (req, res) => {
+	db.Video.find({ user: req.user })
 		.then(dbVideos => res.json(dbVideos))
 		.catch(err => res.json(err));
 });
 
-router.post('/', (req, res) => {
-	const userId = '5e1e6c078bd03303a274ba33';
-	const newVideo = { userId, ...req.body };
+//Create video
+router.post('/', auth, (req, res) => {
+	const newVideo = { user: req.user, ...req.body };
 	db.Video.create(newVideo)
 		.then(dbVideos => res.json(dbVideos))
 		.catch(err => res.json(err));
+	console.log(newVideo);
 });
-
-router.delete('/:id', function(req, res) {
-	const userId = '5e1e6c078bd03303a274ba33';
+router.delete('/:id', auth, (req, res) => {
+	// Make sure user delete only their own video
+	if (dbVideos.user.toString() !== req.user)
+		return res.status(401).json({ msg: 'Not authorized' });
 	db.Video.findByIdAndDelete(req.params.id)
 		.then(dbVideo => res.json(dbVideo))
 		.catch(err => res.json(err));
